pyproject.toml
[project]
name = "odoo-addons-openg2p-registry"
version = "0.0.0"
dependencies = [
    "odoo-addon-g2p_auth_id_oidc @ {root:uri}/g2p_auth_id_oidc",
    "odoo-addon-g2p_auth_oidc @ {root:uri}/g2p_auth_oidc",
    "odoo-addon-g2p_bank @ {root:uri}/g2p_bank",
    "odoo-addon-g2p_bank_rest_api @ {root:uri}/g2p_bank_rest_api",
    "odoo-addon-g2p_change_log @ {root:uri}/g2p_change_log",
    "odoo-addon-g2p_disable_password_login @ {root:uri}/g2p_disable_password_login",
    "odoo-addon-g2p_documents @ {root:uri}/g2p_documents",
    "odoo-addon-g2p_encryption @ {root:uri}/g2p_encryption",
    "odoo-addon-g2p_encryption_keymanager @ {root:uri}/g2p_encryption_keymanager",
    "odoo-addon-g2p_encryption_rest_api @ {root:uri}/g2p_encryption_rest_api",
    "odoo-addon-g2p_enumerator @ {root:uri}/g2p_enumerator",
    "odoo-addon-g2p_mts @ {root:uri}/g2p_mts",
    "odoo-addon-g2p_odk_importer @ {root:uri}/g2p_odk_importer",
    "odoo-addon-g2p_odk_user_mapping @ {root:uri}/g2p_odk_user_mapping",
    "odoo-addon-g2p_openid_vci @ {root:uri}/g2p_openid_vci",
    "odoo-addon-g2p_openid_vci_group @ {root:uri}/g2p_openid_vci_group",
    "odoo-addon-g2p_openid_vci_rest_api @ {root:uri}/g2p_openid_vci_rest_api",
    "odoo-addon-g2p_portal_auth @ {root:uri}/g2p_portal_auth",
    "odoo-addon-g2p_profile_image @ {root:uri}/g2p_profile_image",
    "odoo-addon-g2p_registry_addl_info @ {root:uri}/g2p_registry_addl_info",
    "odoo-addon-g2p_registry_base @ {root:uri}/g2p_registry_base",
    "odoo-addon-g2p_registry_documents @ {root:uri}/g2p_registry_documents",
    "odoo-addon-g2p_registry_encryption @ {root:uri}/g2p_registry_encryption",
    "odoo-addon-g2p_registry_group @ {root:uri}/g2p_registry_group",
    "odoo-addon-g2p_registry_individual @ {root:uri}/g2p_registry_individual",
    "odoo-addon-g2p_registry_membership @ {root:uri}/g2p_registry_membership",
    "odoo-addon-g2p_registry_rest_api @ {root:uri}/g2p_registry_rest_api",
    "odoo-addon-g2p_service_provider_beneficiary_management @ {root:uri}/g2p_service_provider_beneficiary_management",
    "odoo-addon-g2p_service_provider_portal_base @ {root:uri}/g2p_service_provider_portal_base",
    "odoo-addon-g2p_superset_dashboard @ {root:uri}/g2p_superset_dashboard",
    "odoo-addon-mts_connector @ {root:uri}/mts_connector",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
bypass-selection = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
